backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Data Validation
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator

# HTTP Client
httpx==0.27.2

# Async File Operations
aiofiles==24.1.0

# Claude Agent SDK
claude-agent-sdk

# Logging
loguru

# Environment
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.2

# MinIO / S3 (for file uploads)
boto3>=1.26.0
botocore>=1.29.0

# PowerPoint Generation
python-pptx

# YAML parsing for skill frontmatter
pyyaml>=6.0

# File format support (for docx and pdf parsing)
python-docx>=1.1.0  # For .docx file parsing
pdfplumber>=0.10.0  # For .pdf file parsing (recommended, better than PyPDF2)
PyPDF2>=3.0.0  # Alternative PDF parser (fallback if pdfplumber not available)

# Python dependencies for claude-docs skill
# Install with: pip install -r requirements.txt
#
# For agentic tools (Claude Code, etc.):
#   python .claude/skills/claude-docs/scripts/setup_dependencies.py --install-required
#
# Or install manually:
#   pip install -r .claude/skills/claude-docs/requirements.txt

# Core dependencies (required)
requests>=2.28.0
beautifulsoup4>=4.11.0
markdownify>=0.11.6

# NLP libraries (optional but recommended)
# spaCy: Install model after installing package
#   Recommended: python -m spacy download en_core_web_sm
#   For specific version: python -m spacy download en_core_web_sm-3.7.1
#   Check version: python -m spacy info en_core_web_sm
# Note: Model version affects stop word lists - use consistent versions across environments
spacy>=3.7.0
yake>=0.4.8

# Notes:
# - If spaCy is not available, scripts fall back to basic stop words
# - If YAKE is not available, keyword extraction uses heading/content analysis only
# - All scripts work without optional dependencies (with reduced functionality)
# - For docx/pdf support: Install python-docx and pdfplumber (or PyPDF2)
#
# Installation Notes:
# - Pre-built wheels are preferred (no compiler needed) - auto-installation tries these first
# - If wheels aren't available for your Python version, source install may require C++ build tools
# - Python 3.11/3.12 have better pre-built wheel support than newer versions
# - On Windows without build tools: scripts work fine with fallbacks (no action needed)
# - pdfplumber may require additional system dependencies (poppler-utils) on Linux

tavily-python>=0.3.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
coverage==7.6.1
